% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve-sysdeps.R
\name{create_sysreqs_site}
\alias{create_sysreqs_site}
\title{Create a local system requirements site}
\usage{
create_sysreqs_site(outdir = "_api", include_index = TRUE)
}
\arguments{
\item{outdir}{default \code{"_api"}. The path to write the site files to.}

\item{include_index}{default \code{TRUE}. Whether or not to create \code{index.html} files in the site.}
}
\description{
Downloads the system requirement data to a local directory. This can be used to create your own resource to call or serve offline.
}
