% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-fxs.R
\name{fetch_patterns}
\alias{fetch_patterns}
\title{Return list of library regex patterns}
\usage{
fetch_patterns()
}
\description{
Returns a name list containing the regex patterns stored in each rule. The name is the name of the library and the values are the regex patterns used to match the libary.
}
