% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-fxs.R
\name{fetch_pkg}
\alias{fetch_pkg}
\title{Get a packages information}
\usage{
fetch_pkg(pkg, verbose = TRUE)
}
\arguments{
\item{pkg}{package name as a scalar character}

\item{verbose}{default \code{TRUE}. Whether to inform of missing packages.}
}
\description{
Returns the parsed json from \verb{/pkgs/pkg-name.json}.
Containing the package name, version, and system requirements.
}
