% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-fxs.R
\name{fetch_pkg_deps}
\alias{fetch_pkg_deps}
\title{Get a packages system requirements}
\usage{
fetch_pkg_deps(pkg, verbose = TRUE)
}
\arguments{
\item{pkg}{package name as a scalar character}

\item{verbose}{default \code{TRUE}. Whether to inform of missing packages.}
}
\description{
Fetches a package's information and identifies its required system libraries using the patterns stored in each rule. See \code{fetch_patterns()}.
}
