% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-fxs.R
\name{fetch_rules}
\alias{fetch_rules}
\title{Get rules for a given system library}
\usage{
fetch_rules(lib)
}
\arguments{
\item{lib}{name of a system library as scalar character}
}
\description{
Get rules for a given system library
}
