% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-sysdeps.R
\name{get_lib_reqs}
\alias{get_lib_reqs}
\title{Identify system library requirements}
\usage{
get_lib_reqs(lib, os, os_release)
}
\arguments{
\item{lib}{the name of a system library. Must match the name of the library as written in \verb{/rules}.}

\item{os}{the operating system. See \code{\link[=fetch_systems]{fetch_systems()}}.}

\item{os_release}{the version for \code{os}.}
}
\description{
Identify system library requirements
}
\examples{
get_lib_reqs("fftw3", "ubuntu", "18.04")
}
