% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-sysreqs.R
\name{get_pkg_sysreqs}
\alias{get_pkg_sysreqs}
\title{Get system requirements for a package}
\usage{
get_pkg_sysreqs(pkg, os, os_release, verbose = TRUE)
}
\arguments{
\item{pkg}{package name as a scalar character}

\item{os}{the operating system. See \code{\link[=fetch_systems]{fetch_systems()}}.}

\item{os_release}{the version for \code{os}.}

\item{verbose}{default \code{TRUE}. Whether to inform of missing packages.}
}
\value{
A named list for each system requirement identified where each element is a named list for each required system library containing the elements:
\itemize{
\item \code{dependencies}: the required library names
\item \code{pre_installs}: a character vector of commands typically required before installation
\item \code{post_installs}: a character vector of commands typically required after installation
}
}
\description{
From a given R package name, identify it's system requirements.
}
\examples{
get_pkg_sysreqs("igraph", "centos", "8")
}
