% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-sysreqs.R
\name{get_pkgs_sysreqs}
\alias{get_pkgs_sysreqs}
\title{Get system requirements for multiple packages}
\usage{
get_pkgs_sysreqs(pkgs, os, os_release, verbose = FALSE)
}
\arguments{
\item{pkgs}{a character vector of package names.}

\item{os}{the operating system. See \code{\link[=fetch_systems]{fetch_systems()}}.}

\item{os_release}{the version for \code{os}.}

\item{verbose}{default \code{FALSE}. Whether to inform of missing packages.}
}
\value{
A tibble with 4 columns:
\itemize{
\item \code{name}: The system library name.
\item \code{dependencies}: The dependencies for \code{name}.
\item \code{pre_installs}: The recommended pre-installation commands.
\item \code{post_installs}: The recommended post-installation commands.
}
}
\description{
From a vector of R package names, identify their system dependencies.
}
\examples{
get_pkgs_sysreqs(
  c("rJava", "igraph", "rgdal", "spdep"),
  "ubuntu", "16.04"
)
}
