% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serve-sysdeps.R
\name{serve_sysreqs}
\alias{serve_sysreqs}
\title{Serve a local version of the system requirements site}
\usage{
serve_sysreqs(dir = "_api", host = "0.0.0.0", port = 8080)
}
\arguments{
\item{dir}{default \code{"_api"}. The directory to serve}

\item{host}{default localhost. The host to serve on.}

\item{port}{default \code{8080}. The port to listen on.}
}
\description{
To stop serving the site, call \code{\link[=stop_sysreqs]{stop_sysreqs()}}.
}
